Update parser by using regex, this helped me fix some of the bugs and shortcomings of the previous version of the parser

+++ src/main/java/seedu/rainyDay/modules/Ui.java
@@ -102,13 +102,13 @@ public class Ui {
 
     public static void printFinancialStatement(int statementIndex, FinancialStatement currentStatement) {
         String statementName = currentStatement.getDescription();
-        int statementValue = currentStatement.getValue();
+        double statementValue = currentStatement.getValue();
         String statementCategory = currentStatement.getCategory();
         String statementDirection = currentStatement.getFlowSymbol();
 
         String index = String.format("00000%d", statementIndex);
         index = index.substring(index.length()-5);
-        String value = String.format(" %s$%d            ", statementDirection, statementValue);
+        String value = String.format(" %s$%.2f            ", statementDirection, statementValue);
         value = value.substring(0, 12);
         String name = String.format("%s                              ", statementName);
         name = name.substring(0, 30);
